requirements.txt
python-telegram-bot==20.7
pytz==2024.1
kivy==2.2.1
buildozer==1.5.0
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
requests==2.31.0
pyjwt==2.8.0 